Fix SystemError and env lookup
