+++ start.go
@@ -1,32 +1,39 @@
 package main
 
 import (
+	"fmt"
+
 	"github.com/codegangsta/cli"
 	"github.com/opencontainers/runc/libcontainer"
 )
 
 var startCommand = cli.Command{
 	Name:  "start",
-	Usage: "start signals a created container to execute the users defined process",
+	Usage: "start signals a created container to execute the user defined process",
 	ArgsUsage: `<container-id>
 
 Where "<container-id>" is your name for the instance of the container that you
 are starting. The name you provide for the container instance must be unique on
 your host.`,
 	Description: `The start command signals the container to start the user's defined process.`,
-	Action: func(context *cli.Context) {
+	Action: func(context *cli.Context) error {
 		container, err := getContainer(context)
 		if err != nil {
-			fatal(err)
+			return err
 		}
 		status, err := container.Status()
 		if err != nil {
-			fatal(err)
+			return err
 		}
-		if status == libcontainer.Created {
-			if err := container.Signal(libcontainer.InitContinueSignal); err != nil {
-				fatal(err)
-			}
+		switch status {
+		case libcontainer.Created:
+			return container.Signal(libcontainer.InitContinueSignal)
+		case libcontainer.Stopped:
+			return fmt.Errorf("cannot start a container that has run and stopped")
+		case libcontainer.Running:
+			return fmt.Errorf("cannot start an already running container")
+		default:
+			return fmt.Errorf("cannot start a container in the %s state", status)
 		}
 	},
 }
